Sorting out BMI.py

+++ bmi.py
@@ -3,8 +3,8 @@
 
 # Inputting height and weight
 
-height = int (input ("Enter height (cm): "))                # convert input to int, because it's str by default
-weight = int (input ("Enter weight (kg): "))
+height = float (input ("Enter height (cm): "))                # convert input to float, because it's str by default
+weight = float (input ("Enter weight (kg): "))
 
 # Calculating and rounding BMI to 2 decimal places
